API References

API References

Populations

Synapses

Stimuli

Plasticity

Functions

Other types

Helper macros

Populations

SpikingNeuralNetworks.AbstractPopulation — Type

AbstractPopulation

An abstract type representing a population. Any struct inheriting from this type must implement:

Methods

integrate!(p::NeuronModel, param::NeuronModelParam, dt::Float32): Integrates the neuron model over a time step dt using the given parameters.

SpikingNeuralNetworks.BallAndStick — Type

This is a struct representing a spiking neural network model that include two dendrites and a soma based on the adaptive exponential integrate-and-fire model (AdEx)

Fields

t::VIT : tracker of simulation index [0]

param::AdExSoma : Parameters for the AdEx model.

N::Int32 : The number of neurons in the network.

soma_syn::ST : Synapses connected to the soma.

dend_syn::ST : Synapses connected to the dendrites.

d::VDT, d2::VDT : Dendrite structures.

NMDA::NMDAT : Specifies the properties of NMDA (N-methyl-D-aspartate) receptors.

v_s::VFT : Somatic membrane potential.

w_s::VFT : Adaptation variables for each soma.

v_d::VFT , v_d::VFT : Dendritic membrane potential for dendrite 1 and 2.

g_s::MFT , g_d::MFT, g_d2::MFT : Conductance of somatic and dendritic synapses.

h_s::MFT, h_d::MFT, h_d2::MFT : Synaptic gating variables.

fire::VBT : Boolean array indicating which neurons have fired.

after_spike::VFT : Post-spike timing.

postspike::PST : Model for post-spike behavior.

θ::VFT : Individual neuron firing thresholds.

records::Dict : A dictionary to store simulation results.

Δv::VFT , Δv_temp::VFT : Variables to hold temporary voltage changes.

cs::VFT , is::VFT : Temporary variables for currents.

SpikingNeuralNetworks.HH — Type

Hodgkin–Huxley Neuron

SpikingNeuralNetworks.IF — Type

[Integrate-And-Fire Neuron](https://neuronaldynamics.epfl.ch/online/Ch1.S3.html)

SpikingNeuralNetworks.IZ — Type

Izhikevich Neuron

SpikingNeuralNetworks.NoisyIF — Type

Noisy Integrate-And-Fire Neuron

SpikingNeuralNetworks.Poisson — Type

Poisson Neuron

SpikingNeuralNetworks.Rate — Type

Rate Neuron

SpikingNeuralNetworks.Tripod — Type

This is a struct representing a spiking neural network model that include two dendrites and a soma based on the adaptive exponential integrate-and-fire model (AdEx)

Fields

t::VIT : tracker of simulation index [0]

param::AdExSoma : Parameters for the AdEx model.

N::Int32 : The number of neurons in the network.

soma_syn::ST : Synapses connected to the soma.

dend_syn::ST : Synapses connected to the dendrites.

d1::VDT, d2::VDT : Dendrite structures.

NMDA::NMDAT : Specifies the properties of NMDA (N-methyl-D-aspartate) receptors.

gax1::VFT, gax2::VFT : Axial conductance (reciprocal of axial resistance) for dendrite 1 and 2 respectively.

cd1::VFT, cd2::VFT : Capacitance for dendrite 1 and 2.

gm1::VFT, gm2::VFT : Membrane conductance for dendrite 1 and 2.

v_s::VFT : Somatic membrane potential.

w_s::VFT : Adaptation variables for each soma.

v_d1::VFT , v_d2::VFT : Dendritic membrane potential for dendrite 1 and 2.

g_s::MFT , g_d1::MFT, g_d2::MFT : Conductance of somatic and dendritic synapses.

h_s::MFT, h_d1::MFT, h_d2::MFT : Synaptic gating variables.

fire::VBT : Boolean array indicating which neurons have fired.

after_spike::VFT : Post-spike timing.

postspike::PST : Model for post-spike behavior.

θ::VFT : Individual neuron firing thresholds.

records::Dict : A dictionary to store simulation results.

Δv::VFT , Δv_temp::VFT : Variables to hold temporary voltage changes.

cs::VFT , is::VFT : Temporary variables for currents.

Synapses

SpikingNeuralNetworks.AbstractConnection — Type

AbstractConnection

An abstract type representing a connection. Any struct inheriting from this type must implement:

Methods

forward!(c::Synapse, param::SynapseParameter): Propagates the signal through the synapse.

plasticity!(c::Synapse, param::SynapseParameter, dt::Float32, T::Time): Updates the synapse parameters based on plasticity rules.

SpikingNeuralNetworks.AbstractNormalization — Type

AbstractNormalization <: AbstractConnection

An abstract type representing a normalization connection.

SpikingNeuralNetworks.AbstractSparseSynapse — Type

AbstractSparseSynapse <: AbstractConnection

An abstract type representing a sparse synapse connection.

SpikingNeuralNetworks.AggregateScaling — Method

SynapseNormalization(N; param, kwargs...)

Constructor function for the SynapseNormalization struct.

N: The number of synapses.

param: Normalization parameter, can be either MultiplicativeNorm or AdditiveNorm.

kwargs: Other optional parameters.

Returns a SynapseNormalization object with the specified parameters.

SpikingNeuralNetworks.FLSparseSynapse — Type

Force Learning Sparse Synapse

SpikingNeuralNetworks.FLSynapse — Type

Force Learning Full Synapse

SpikingNeuralNetworks.PINningSparseSynapse — Type

PINing Sparse Synapse

SpikingNeuralNetworks.PINningSynapse — Type

PINing Sparse Synapse

SpikingNeuralNetworks.RateSynapse — Type

Rate Synapse

SpikingNeuralNetworks.SpikeRateSynapse — Type

Rate Synapse

SpikingNeuralNetworks.SynapseNormalization — Type

SynapseNormalization{VFT = Vector{Float32}, VIT = Vector{Int32}, MFT = Matrix{Float32}}

A struct that holds parameters for synapse normalization, including:

param: Normalization parameter, can be either MultiplicativeNorm or AdditiveNorm.

t: A vector of integer values representing time points.

W0: A vector of initial weights before simulation.

W1: A vector of weights during the simulation.

μ: A vector of mean synaptic weights.

records: A dictionary for storing additional data.

SpikingNeuralNetworks.SynapseNormalization — Method

SynapseNormalization(N; param, kwargs...)

Constructor function for the SynapseNormalization struct.

N: The number of synapses.

param: Normalization parameter, can be either MultiplicativeNorm or AdditiveNorm.

kwargs: Other optional parameters.

Returns a SynapseNormalization object with the specified parameters.

Stimuli

SpikingNeuralNetworks.AbstractStimulus — Type

AbstractStimulus

An abstract type representing a stimulus. Any struct inheriting from this type must implement:

Methods

stimulate!(p::Stimulus, param::StimulusParameter, time::Time, dt::Float32): Applies the stimulus to the population.

SpikingNeuralNetworks.BalancedStimulus — Method

BalancedStimulus(post::T, sym::Symbol, r::Union{Function, Float32}, neurons=[]; N_pre::Int=50, p_post::R=0.05f0, μ::R=1.f0, param=BalancedParameter()) where {T <: AbstractPopulation, R <: Number}

Constructs a BalancedStimulus object for a spiking neural network.

Arguments

post::T: The target population for the stimulus.

sym::Symbol: The symbol representing the synaptic conductance or current.

r::Union{Function, Float32}: The firing rate of the stimulus. Can be a constant value or a function of time.

neurons=[]: The indices of the neuronsin the target population that receive the stimulus. If empty, neuronsare randomly selected based on the probability p_post.

N::Int=200: The number of Balanced neurons neurons.

N_pre::Int=5: The number of presynaptic connected.

p_post::R=0.05f0: The probability of connection between presynaptic and postsynaptic neurons.

μ::R=1.f0: The scaling factor for the synaptic weights.

param=BalancedParameter(): The parameters for the Balanced distribution.

Returns

A BalancedStimulus object.

Plasticity

SpikingNeuralNetworks.STDPGerstner — Type

Gerstner, W., Kempter, R., van Hemmen, J. L., & Wagner, H. (1996). A neuronal learning rule for sub-millisecond temporal coding. Nature, 383(6595), 76–78. https://doi.org/10.1038/383076a0

SpikingNeuralNetworks.STDPMexicanHat — Type

STDPMexicanHat{FT = Float32}

The STDP is defined such that integral of the kernel is zero. The STDP kernel is defined as:

`` A x * exp(-x/sqrt(2)) ``

where   ``A`` is the learning rate for post and pre-synaptic spikes, respectively, and ``x`` is the difference between the post and pre-synaptic traces.

SpikingNeuralNetworks.STDPSymmetric — Type

STDPSymmetric{FT = Float32}

Symmetric STDP rules described in:
`Structured stabilization in recurrent neural circuits through inhibitory synaptic plasticity` 
by Festa, D., Cusseddu, C, and Gjorgjieva, J. (2024).

The STDP is defined such that integral of the kernel is zero. The STDP kernel is defined as:

`` (rac{A_{post}}{1/	au_{post}} * exp(-t/	au_{post} - rac{A_{pre}}{	au_pre} * exp(-t/	au_{pre}) ``

where ``A_{post}`` and ``A_{pre}`` are the learning rates for post and pre-synaptic spikes, respectively, and ``	au_{post}`` and ``	au_{pre}`` are the time constants for post and pre-synaptic traces, respectively.

SpikingNeuralNetworks.vSTDPParameter — Type

vSTDPParameter{FT = Float32} <: SpikingSynapseParameter

Contains parameters for the voltage-dependent Spike-Timing Dependent Plasticity (vSTDP) model.

Fields

A_LTD::FT: Long Term Depression learning rate.

A_LTP::FT: Long Term Potentiation learning rate.

θ_LTD::FT: LTD threshold.

θ_LTP::FT: LTP threshold.

τu::FT: Time constant for the pre-synaptic spike trace.

τv::FT: Time constant for the post-synaptic membrane trace.

τx::FT: Time constant for the variable x.

Wmax::FT: Maximum synaptic weight.

Wmin::FT: Minimum synaptic weight.

active::Vector{Bool}: Flag to activate or deactivate the plasticity rule.

Functions

SpikingNeuralNetworks.CV_isi2 — Method

CV_isi2(intervals::Vector{Float32})

Return the local coefficient of variation of the interspike intervals
Holt, G. R., Softky, W. R., Koch, C., & Douglas, R. J. (1996). Comparison of discharge variability in vitro and in vivo in cat visual cortex neurons. Journal of Neurophysiology, 75(5), 1806–1814. https://doi.org/10.1152/jn.1996.75.5.1806

SpikingNeuralNetworks.C_mem — Method

C_mem(;Cd=Cd,d=d,l=l)
Capacitance of a cylinder of length l and diameter d
return Capacitance in pF

SpikingNeuralNetworks.G_axial — Method

G_axial(;Ri=Ri,d=d,l=l)
Axial conductance of a cylinder of length l and diameter d
return Conductance in nS

SpikingNeuralNetworks.G_mem — Method

G_mem(;Rd=Rd,d=d,l=l)
Membrane conductance of a cylinder of length l and diameter d
return Conductance in nS

SpikingNeuralNetworks.alpha_function — Method

alpha_function(t::T; t0::T, τ::T) where T <: AbstractFloat

Alpha function for convolution of spiketimes. Evaluate the alpha function at time t, with time of peak t0 and time constant τ.

SpikingNeuralNetworks.bin_spiketimes — Method

bin_spiketimes(spiketimes, interval, sr)

Given a list of spike times spiketimes, an interval [start, end], and a sampling rate sr, this function counts the number of spikes that fall within each time bin of width 1/sr within the interval. The function returns a sparse matrix count containing the spike counts for each bin, and an array r containing the time points corresponding to the center of each bin.

Arguments

spiketimes: A 1-dimensional array of spike times.

interval: A 2-element array specifying the start and end times of the interval.

sr: The sampling rate, i.e., the number of time bins per second.

Returns

count: A sparse matrix containing the spike counts for each time bin.

r: An array of time points corresponding to the center of each time bin.

SpikingNeuralNetworks.clear_monitor! — Method

clear_monitor!(obj)

Clears all the records of a given object.

SpikingNeuralNetworks.clear_records! — Method

clear_records!(objs::AbstractArray)

Clears the records of multiple objects.

SpikingNeuralNetworks.clear_records! — Method

clear_records!(obj, sym::Symbol)

Clears the records of a given object for a specific symbol.

SpikingNeuralNetworks.clear_records! — Method

clear_records!(obj)

Clears all the records of a given object.

SpikingNeuralNetworks.compute_connections — Method

compute_connections(pre::Symbol, post::Symbol, points; dc, pl, ϵ, grid_size, conn)

Compute the connections between two populations of neurons based on their spatial distance. This function will assign connections with probability p_short for short-range connections and p_long for long-range connections. The weights of the connections are determined by the μ parameter in the conn named tuple. The function uses a periodic boundary condition to calculate distances in a 2D grid. The total number of connections per the post-synaptic neuron is: ϵ * Npre * pshort + (1 - ϵ) * Npre * plong.

Arguments

pre::Symbol: The symbol representing the pre-synaptic population.

post::Symbol: The symbol representing the post-synaptic population.

points::NamedTuple: A named tuple containing the spatial points for each population.

dc::Float64: The critical distance for short-range connections.

pl::Float64: The probability of long-range connections.

ϵ::Float64: The scaling factor for connection probabilities.

grid_size::Float64: The size of the grid.

conn::NamedTuple: A named tuple containing connection parameters, including p and μ.

Returns

P::Matrix{Bool}: A matrix indicating the presence of connections.

W::Matrix{Float32}: A matrix containing the weights of the connections.

SpikingNeuralNetworks.compute_covariance_density — Method

compute_covariance_density(t_post, t_pre, T; τ=200ms, sr=50Hz)

Compute the covariance density of spike trains t_post and t_pre over a time interval T. The function returns the covariance density vectors for positive and negative time lags.

Arguments

spike_times1:: Vector{Float32}: A vector of spike times for the first neuron.

spike_times2:: Vector{Float32}: A vector of spike times for the second neuron.

bin_width:: Float32: The width of the time bins in milliseconds.

max_lag:: Float32: The maximum time lag in milliseconds.

Returns

lags:: Vector{Float32}: The time lags in milliseconds.

covariance_density:: Vector{Float32}: The covariance density for each time lag.

SpikingNeuralNetworks.compute_cross_correlogram — Function

autocor(spiketimes::Spiketimes; interval = 0:1:1000)

Calculate the cross-correlation of two spike trains.

Arguments

spike_times1: A vector of spike times for the first neuron.

spike_times2: A vector of spike times for the second neuron.

bin_width: The width of the time bins in milliseconds.

max_lag: The maximum time lag in milliseconds.

Returns

lags: The time lags in milliseconds.

auto_corr: The auto-correlation for each time lag.

SpikingNeuralNetworks.create_spatial_structure — Method

create_spatial_structure(config)

Create a 2D spatial structure and dispose N points for each population.

Arguments

config::Dict: A dictionary containing configuration parameters, including projections and Npop.

Returns

Pops::NamedTuple: A named tuple containing the spatial points for each population.

SpikingNeuralNetworks.extract_items — Method

extract_items(root::Symbol, container; pop::Dict{Symbol,Any}, syn::Dict{Symbol, Any}, stim::Dict{Symbol,Any})

Extracts items from a container and adds them to the corresponding dictionaries based on their type.

Arguments

root::Symbol: The root symbol for the items being extracted.

container: The container from which to extract items.

pop::Dict{Symbol,Any}: The dictionary to store population items.

syn::Dict{Symbol, Any}: The dictionary to store synapse items.

stim::Dict{Symbol,Any}: The dictionary to store stimulus items.

Returns

true: Always returns true.

Details

If the type of the item in the container is AbstractPopulation, it is added to the pop dictionary.

If the type of the item in the container is AbstractConnection, it is added to the syn dictionary.

If the type of the item in the container is AbstractStimulus, it is added to the stim dictionary.

If the type of the item in the container is none of the above, the function is recursively called to extract items from the nested container.

SpikingNeuralNetworks.firing_rate — Method

firing_rate(
    spiketimes::Spiketimes,
    interval::AbstractVector = [],
    sampling = 20,
    τ = 25,
    ttf = -1,
    tt0 = -1,
    cache = true,
    pop::Union{Symbol,Vector{Int}}= :ALL,
)

Calculate the firing rates for a population or an individual neuron.

Arguments

spiketimes: Spiketimes object.

interval: Time interval vector (default is an empty vector).

sampling: Sampling rate (default is 20ms).

τ: Time constant for convolution (default is 25ms).

ttf: Final time point (default is -1, which means until the end of the simulation time).

tt0: Initial time point (default is -1, which means from the start of the simulation time based on the sampling rate).

cache: If true, uses cached data (default is true).

pop: Either :ALL for all populations or a Vector of Integers specifying specific neuron indices. Default is :ALL.

Returns

A tuple containing:

rates: A vector of firing rates for each neuron in the chosen population.

interval: The time interval over which the firing rates were calculated.

Examples

SpikingNeuralNetworks.firing_rate_average — Method

firing_rate_average(P; dt=0.1ms)

Calculates and returns the average firing rates of neurons in a network.

Arguments:

P: A structure containing neural data, with a key :fire in its records field which stores spike information for each neuron.

dt: An optional parameter specifying the time interval (default is 0.1ms).

Returns:

An array of floating point values representing the average firing rate for each neuron.

Usage:# Notes:

Each row of P.records[:fire] represents a neuron, and each column represents a time point. The value in a cell indicates whether that neuron has fired at that time point (non-zero value means it has fired). The firing rate of a neuron is calculated as the total number of spikes divided by the total time span.

SpikingNeuralNetworks.gaussian_kernel — Method

gaussian_kernel(σ::Float64, length::Int)

Create a Gaussian kernel with standard deviation σ and specified length.

Arguments

σ: Standard deviation of the Gaussian kernel.

length: Length of the kernel.

Returns

kernel: A vector representing the Gaussian kernel.

SpikingNeuralNetworks.gaussian_kernel_estimate — Method

gaussian_kernel_estimate(support_vector::Vector{Float64}, σ::Float64, length::Int)

Apply a Gaussian kernel estimate to a support vector with closed boundary conditions.

Arguments

support_vector: The input support vector.

σ: Standard deviation of the Gaussian kernel.

length: Length of the kernel.

Returns

estimated_vector: The estimated vector after applying the Gaussian kernel.

SpikingNeuralNetworks.get_dt — Method

get_dt(T::Time)

Get the time step size.

Arguments

T::Time: The Time object.

Returns

Float32: The time step size.

SpikingNeuralNetworks.get_interval — Method

get_interval(T::Time)

Get the time interval from 0 to the current time.

Arguments

T::Time: The Time object.

Returns

StepRange{Float32}: The time interval.

SpikingNeuralNetworks.get_step — Method

get_step(T::Time)

Get the current time step.

Arguments

T::Time: The Time object.

Returns

Float32: The current time step.

SpikingNeuralNetworks.get_time — Method

get_time(T::Time)

Get the current time.

Arguments

T::Time: The Time object.

Returns

Float32: The current time.

SpikingNeuralNetworks.getrecord — Method

getrecord(p, sym)

Returns the recorded values for a given object and symbol. If the symbol is not found in the object's records, it checks the records of the object's plasticity and returns the values for the matching symbol.

SpikingNeuralNetworks.getvariable — Function

getvariable(obj, key, id=nothing)

Returns the recorded values for a given object and key. If an id is provided, returns the recorded values for that specific id.

SpikingNeuralNetworks.graph — Method

graph(model)

Generate a graph representation of the model.

Arguments

model: The model to generate the graph from.

Returns

A MetaGraphs.MetaDiGraph object representing the graph.

Details

Each vertex represents either a population ('pop'), a normalization synapse ('norm'), or a stimulus pre-target ('pre'). Its metadata includes:

name: Actual name of the population, 'norm' for a SynapseNormalization, or the pre-target's name for a stimulus.

id: Identifier of the population, SynapseNormalization, or stimulus.

key: Key from the original 'pop', 'syn', or 'stim' dictionary in the model.

Each edge represents a synaptic connection or a stimulus. Its metadata includes:

type: Type of the edge, ':firetog' for SpikingSynapse, ':norm' for SynapseNormalization, or ':stim' for a stimulus.

name: Name of the edge, formatted as "fromvertexname to tovertexname".

key: Key from the original 'syn' or 'stim' dictionary in the model.

id: Identifier of the synapse or stimulus.

The function iterates over the populations, synapses, and stimuli in the model.

AbstractPopulation items are added as vertices.

For each connection it checks the type of the synapse and adds an edge between the pre-synaptic population and the post-synaptic population. - SNN.SpikingSynapse: the edge represents a connection from the firing population to the receiving population. - SNN.SynapseNormalization: the edge represents a normalization of synapses between populations. - SNN.PoissonStimulus: the edge represents a stimulus from the pre-synaptic population to the post-synaptic population.

For each stimulus, it adds a vertex to the graph representing an implicit pre-synaptic population [:fire] an edge between it and the post-synaptic population [:post].

Returns a MetaGraphs.MetaDiGraph where:

Errors

Throws ArgumentError when the synapse type is neither SNN.SpikingSynapse nor SNN.SynapseNormalization.

SpikingNeuralNetworks.interpolated_record — Method

interpolated_record(p, sym)

Returns the recording with interpolated time values and the extrema of the recorded time points.

N.B. 
----
The element can be accessed at whichever time point by using the index of the array. The time point must be within the range of the recorded time points, in r_v.

SpikingNeuralNetworks.interval_standard_spikes — Method

interval_standard_spikes(spiketimes, interval)

Standardize the spiketimes to the interval [0, interval_duration]. Return a copy of the 'Spiketimes' vector.

SpikingNeuralNetworks.is_attractor_state — Method

is_attractor_state(spiketimes::Spiketimes, interval::AbstractVector, N::Int)

Check if the network is in an attractor state by verifying that the average firing rate over the last N seconds of the simulation is a unimodal distribution.

Arguments

spiketimes: A vector of vectors containing the spike times of each neuron.

interval: The time interval over which to compute the firing rate.

N: The number of seconds over which to check the unimodality of the firing rate distribution.

Returns

is_attractor: A boolean indicating whether the network is in an attractor state.

SpikingNeuralNetworks.linear_network — Method

linear_network(N, σ_w=0.38, w_max=2.0)

Create a linear network with Gaussian-shaped connections.

Arguments

N::Int: The number of neurons.

σ_w::Float64: The standard deviation of the Gaussian distribution.

w_max::Float64: The maximum weight.

Returns

W::Matrix{Float32}: A matrix containing the weights of the connections.

SpikingNeuralNetworks.merge_models — Method

merge_models(kwargs...; syn=nothing, pop=nothing)

Merge multiple models into a single model.

Arguments

kwargs...: List of kwarg elements, i.e., dictionary or named tuples, containing the models to be merged.

if kwarg has elements with :pop and :syn entries, the function copies them into the merged model.

if kwarg has no :pop and :syn entries, the function iterates over all the elements contained in kwarg and merge them into the model.

syn: Optional dictionary of synapses to be merged.

pop: Optional dictionary of populations to be merged.

stim: Optional dictionary of stimuli to be merged.

Returns

A tuple (pop, syn) representing the merged populations and synapses.

Details

This function takes in multiple models represented as keyword arguments and merges them into a single model. The models can be specified using the pop and syn fields in the keyword arguments. If the pop and syn fields are not present, the function expects the keyword arguments to have elements with :pop or :syn fields.

The merged populations and synapses are stored in dictionaries populations and synapses, respectively. The function performs type assertions to ensure that the elements being merged are of the correct types (AbstractPopulation for populations and AbstractConnection for synapses).

If syn and/or pop and/or stim arguments are provided, they are merged into the respective dictionaries.

Example

SpikingNeuralNetworks.merge_spiketimes — Method

merge_spiketimes(spikes::Vector{Spiketimes}; )

Merge spiketimes from different simulations. 
This function is not thread safe, it is not recommended to use it in parallel.
Parameters
----------
spikes: Vector{Spiketimes}
    Vector of spiketimes from different simulations
Return
------
neurons: Spiketimes
    Single vector of spiketimes

SpikingNeuralNetworks.monitor! — Method

monitor!(objs::Array, keys)

Function called when more than one object is given, which then calls the above monitor function for each object

SpikingNeuralNetworks.monitor! — Method

Initialize dictionary records for the given object, by assigning empty vectors to the given keys

Arguments

obj: An object whose variables will be monitored

keys: The variables to be monitored

SpikingNeuralNetworks.neurons_outside_area — Method

neurons_outside_area(points, center, distance, grid_size)

Find the indices of neurons outside a specified area around a center point.

Arguments

points::Vector{Vector{Float64}}: The coordinates of the neurons.

center::Vector{Float64}: The coordinates of the center point.

distance::Float64: The minimum distance from the center point.

grid_size::Float64: The size of the grid.

Returns

indices::Vector{Int}: The indices of neurons outside the specified area.

SpikingNeuralNetworks.neurons_within_area — Method

neurons_within_area(points, center, distance, grid_size)

Find the indices of neurons within a specified area around a center point.

Arguments

points::Vector{Vector{Float64}}: The coordinates of the neurons.

center::Vector{Float64}: The coordinates of the center point.

distance::Float64: The maximum distance from the center point.

grid_size::Float64: The size of the grid.

Returns

indices::Vector{Int}: The indices of neurons within the specified area.

SpikingNeuralNetworks.norm_synapse — Method

Calculate the normalization factor for a synapse given rise and decay time constants.

Arguments

τr: Rise time constant

τd: Decay time constant

Returns

Float32: The normalization factor

SpikingNeuralNetworks.norm_synapse — Method

Calculate the normalization factor for a synapse.

Arguments

synapse::Receptor: The receptor for which to calculate the normalization factor

Returns

Float32: The normalization factor

SpikingNeuralNetworks.periodic_distance — Method

periodic_distance(point1, point2, grid_size)

Calculate the periodic distance between two points in a 2D grid.

Arguments

point1::Vector{Float64}: The coordinates of the first point.

point2::Vector{Float64}: The coordinates of the second point.

grid_size::Float64: The size of the grid.

Returns

distance::Float64: The periodic distance between the two points.

SpikingNeuralNetworks.plasticity! — Method

plasticity!(c::SynapseNormalization, param::AdditiveNorm, dt::Float32)

Updates the synaptic weights using additive or multiplicative normalization (operator). This function calculates the rate of change μ as the difference between initial weight W0 and the current weight W1, normalized by W1. The weights are updated at intervals specified by time constant τ.

Arguments

c: An instance of SynapseNormalization.

param: An instance of AdditiveNorm.

dt: Simulation time step.

SpikingNeuralNetworks.plasticity! — Method

plasticity!(c::AbstractSparseSynapse, param::iSTDPRate, dt::Float32)

Performs the synaptic plasticity calculation based on the inihibitory spike-timing dependent plasticity (iSTDP) model from Vogels (2011) adapted to control the membrane potential. The function updates synaptic weights W of each synapse in the network according to the firing status of pre and post-synaptic neurons. This is an in-place operation that modifies the input AbstractSparseSynapse object c.

Arguments

c::AbstractSparseSynapse: The current spiking synapse object which contains data structures to represent the synapse network.

param::iSTDPParameterVoltage: Parameters needed for the iSTDP model, including learning rate η, target membrane potenital v0, STDP time constant τy, maximal and minimal synaptic weight (Wmax and Wmin).

dt::Float32: The time step for the numerical integration.

Algorithm

For each pre-synaptic neuron, if it fires, it reduces the synaptic weight by an amount proportional to the difference between the target membrane potential and the actual one.

For each pre-synaptic neuron, if it fires, increases the inhibitory term, otherwise the inhibitory term decays exponentially over time with a time constant τy.

For each post-synaptic neuron, if it fires, it increases the synaptic weight by an amount proportional to the pre-synaptic trace and increases the excitatory term, otherwise the excitatory term decays exponentially over time with a time constant τy.

The synaptic weights are bounded by Wmin and Wmax.

SpikingNeuralNetworks.plasticity! — Method

plasticity!(c::AbstractSparseSynapse, param::iSTDPRate, dt::Float32)

Performs the synaptic plasticity calculation based on the inihibitory spike-timing dependent plasticity (iSTDP) model from Vogels (2011). The function updates synaptic weights W of each synapse in the network according to the firing status of pre and post-synaptic neurons. This is an in-place operation that modifies the input AbstractSparseSynapse object c.

Arguments

c::AbstractSparseSynapse: The current spiking synapse object which contains data structures to represent the synapse network.

param::iSTDPRate: Parameters needed for the iSTDP model, including learning rate η, target rate r, STDP time constant τy, maximal and minimal synaptic weight (Wmax and Wmin).

dt::Float32: The time step for the numerical integration.

Algorithm

For each pre-synaptic neuron, if it fires, it reduces the synaptic weight by an amount proportional to the difference between the target rate and the actual rate and increases the inhibitory term, otherwise the inhibitory term decays exponentially over time with a time constant τy.

For each post-synaptic neuron, if it fires, it increases the synaptic weight by an amount proportional to the pre-synaptic trace and increases the excitatory term, otherwise the excitatory term decays exponentially over time with a time constant τy.

The synaptic weights are bounded by Wmin and Wmax.

SpikingNeuralNetworks.plasticity! — Method

plasticity!(c::SynapseNormalization, param::AdditiveNorm, dt::Float32)

Updates the synaptic weights using additive or multiplicative normalization (operator). This function calculates the rate of change μ as the difference between initial weight W0 and the current weight W1, normalized by W1. The weights are updated at intervals specified by time constant τ.

Arguments

c: An instance of SynapseNormalization.

param: An instance of AdditiveNorm.

dt: Simulation time step.

SpikingNeuralNetworks.population_indices — Method

population_indices(P, type = "ˆ")

Given a dictionary P containing population names as keys and population objects as values, this function returns a named tuple indices that maps each population name to a range of indices. The range represents the indices of the neurons belonging to that population.

Arguments

P: A dictionary containing population names as keys and population objects as values.

type: A string specifying the type of population to consider. Only population names that contain the specified type will be included in the output. Defaults to "ˆ".

Returns

A named tuple indices where each population name is mapped to a range of indices.

SpikingNeuralNetworks.print_model — Function

print_model(model)

Prints the details of the given model. The model is expected to have three components: pop (populations), syn (synapses), and stim (stimuli).

The function displays a graph representation of the model, followed by detailed information about each component.

Arguments

model: The model containing populations, synapses, and stimuli to be printed.

Outputs

Prints the graph of the model, along with the name, key, type, and parameters of each component in the populations, synapses, and stimuli.

Exception

Raises an assertion error if any component in the populations is not a subtype of SNN.AbstractPopulation, if any component in the synapses is not a subtype of SNN.AbstractConnection, or if any component in the stimuli is not a subtype of SNN.AbstractStimulus.

SpikingNeuralNetworks.print_summary — Method

print_summary(p)

Prints a summary of the given element.

SpikingNeuralNetworks.record_fire! — Method

record_fire!(obj::PT, T::Time, indices::Dict{Symbol,Vector{Int}}) where {PT <: Union{AbstractPopulation, AbstractStimulus}}

Record the firing activity of the obj object into the obj.records[:fire] array.

Arguments

obj::PT: The object to record the firing activity from.

T::Time: The time at which the recording is happening.

indices::Dict{Symbol,Vector{Int}}: A dictionary containing indices for each variable to record.

SpikingNeuralNetworks.record_sym! — Method

record_sym!(obj, key::Symbol, T::Time, indices::Dict{Symbol,Vector{Int}})

Record the variable key of the obj object into the obj.records[key] array.

Arguments

obj: The object to record the variable from.

key::Symbol: The key of the variable to record.

T::Time: The time at which the recording is happening.

indices::Dict{Symbol,Vector{Int}}: A dictionary containing indices for each variable to record.

SpikingNeuralNetworks.relative_time! — Method

relative_time(spiketimes::Spiketimes, start_time)

Return the spiketimes relative to the start_time of the interval

SpikingNeuralNetworks.sample_spikes — Method

sample_spikes(N, rate::Vector, interval::R; dt=0.125f0) where {R <: AbstractRange}

Generate sample spike times for N neurons from a rate vector. The function generates spike times for each neuron based on the rate vector and the time interval. The spike times are generated such that during the interval of the rate, the number of spikes is Poisson distributed with the rate.

Arguments

N: The number of neurons.

rate::Vector: The vector with the rate to be sampled.

interval::R: The time interval over which the rate is recorded.

dt=0.125f0: The time step size.

Returns

An array of spike times for each neuron.

SpikingNeuralNetworks.sim! — Method

sim!(
    P::Vector{TN},
    C::Vector{TS};
    dt = 0.1f0,
    duration = 10.0f0,
    pbar = false,
) where {TN <: AbstractPopulation, TS<:AbstractConnection }

Simulates the spiking neural network for a specified duration by repeatedly calling sim! function.

Arguments

P::Vector{TN}: Vector of neurons in the network.

C::Vector{TS}: Vector of synapses in the network.

dt::Float32: Time step for the simulation. Default value is 0.1f0.

duration::Float32: Duration of the simulation. Default value is 10.0f0.

pbar::Bool: Flag indicating whether to display a progress bar during the simulation. Default value is false.

Details

The function creates a range of time steps from 0.0f0 to duration-dt with a step size of dt.

If pbar is true, the function creates a progress bar using the ProgressBar function with the time step range. Otherwise, it uses the time step range directly.

The function iterates over the time steps and calls the sim! function with P, C, and dt.

SpikingNeuralNetworks.spikecount — Method

spikecount(model, Trange, neurons)

Return the total number of spikes of the neurons in the selected interval

SpikingNeuralNetworks.spikes_in_interval — Function

spikesininterval(spiketimes::Spiketimes, interval::AbstractRange)

Return the spiketimes in the selected interval

Arguments

spiketimes::Spiketimes: Vector with each neuron spiketime interval: 2 dimensional array with the start and end of the interval

SpikingNeuralNetworks.spiketimes — Method

spiketimes(Ps; kwargs...)

Return the spiketimes of each population in single vector of Spiketimes.

SpikingNeuralNetworks.spiketimes — Method

spiketimes(p, interval = nothing, indices = nothing)

Compute the spike times of a population.

Arguments:

p: The network parameters.

interval: The time interval within which to compute the spike times. If nothing, the interval is set to (0, firing_time[end]).

Returns:

spiketimes: A vector of vectors containing the spike times of each neuron.

SpikingNeuralNetworks.spiketimes_from_bool — Method

spiketimes_from_bool(P, τ; dt = 0.1ms)

This function takes in the records of a neural population P and time constant τ to calculate spike times for each neuron.

Arguments

P: A data structure containing the recorded data of a neuronal population.

τ: A time constant parameter.

Keyword Arguments

dt: The time step used for the simulation, defaults to 0.1 milliseconds.

Returns

spiketimes: An object of type SNN.Spiketimes which contains the calculated spike times of each neuron.

Examples

julia
spiketimes = spike_times(population_records, time_constant)

SpikingNeuralNetworks.spiketimes_split — Method

spiketimes_split(Ps; kwargs...)

Return the spiketimes of each population in a vector of Spiketimes.

SpikingNeuralNetworks.st_order — Method

st_order(spiketimes::Spiketimes)

SpikingNeuralNetworks.stimulate! — Method

stimulate!(p::BalancedStimulus, param::BalancedParameter, time::Time, dt::Float32)

Generate a Balanced stimulus for a postsynaptic population.

SpikingNeuralNetworks.subpopulations — Function

subpopulations(stim)

Extracts the names and the neuron ids projected from a given set of stimuli.

Arguments

stim: A dictionary containing stimulus information.

Returns

names: A vector of strings representing the names of the subpopulations.

pops: A vector of arrays representing the populations of the subpopulations.

Example

SpikingNeuralNetworks.synapsearray — Function

Convert a Synapse to a SynapseArray.

Arguments

syn::Synapse: The Synapse object

indices::Vector: Optional vector of indices to include in the SynapseArray

Returns

SynapseArray: The SynapseArray object

SpikingNeuralNetworks.synapsearray — Method

Return the SynapseArray as is.

Arguments

syn::SynapseArray: The SynapseArray object

Returns

SynapseArray: The SynapseArray object

SpikingNeuralNetworks.train! — Method

train!(
    P::Vector{TN},
    C::Vector{TS};
    dt = 0.1ms,
    duration = 10ms,
) where {TN <: AbstractPopulation, TS<:AbstractConnection }

Trains the spiking neural network for a specified duration by repeatedly calling train! function.

Arguments

P::Vector{TN}: Vector of neurons in the network.

C::Vector{TS}: Vector of synapses in the network.

dt::Float32: Time step for the training. Default value is 0.1ms.

duration::Float32: Duration of the training. Default value is 10ms.

Details

The function converts dt to Float32 if it is not already.

The function creates a progress bar using the ProgressBar function with a range of time steps from 0.0f0 to duration-dt with a step size of dt.

The function iterates over the time steps and calls the train! function with P, C, and dt.

SpikingNeuralNetworks.update_time! — Method

update_time!(T::Time, dt::Float32)

Update the current time and time step.

Arguments

T::Time: The Time object.

dt::Float32: The time step size.

SpikingNeuralNetworks.α_synapse — Method

Calculate the alpha factor for a synapse given rise and decay time constants.

Arguments

τr: Rise time constant

τd: Decay time constant

Returns

Float32: The alpha factor

Other types

SpikingNeuralNetworks.AbstractConnectionParameter — Type

AbstractConnectionParameter <: AbstractParameter

An abstract type representing a connection parameter.

SpikingNeuralNetworks.AbstractParameter — Type

AbstractParameter

An abstract type representing a parameter.

SpikingNeuralNetworks.AbstractPopulationParameter — Type

AbstractPopulationParameter <: AbstractParameter

An abstract type representing a population parameter.

SpikingNeuralNetworks.AbstractStimulusParameter — Type

AbstractStimulusParameter <: AbstractParameter

An abstract type representing a stimulus parameter.

SpikingNeuralNetworks.AdExSoma — Type

AdExSoma

An implementation of the Adaptive Exponential Integrate-and-Fire (AdEx) model, adapted for a Tripod neuron.

Fields

C::FT = 281pF: Membrane capacitance.

gl::FT = 40nS: Leak conductance.

R::FT = nS / gl * GΩ: Total membrane resistance.

τm::FT = C / gl: Membrane time constant.

Er::FT = -70.6mV: Resting potential.

Vr::FT = -55.6mV: Reset potential.

Vt::FT = -50.4mV: Rheobase threshold.

ΔT::FT = 2mV: Slope factor.

τw::FT = 144ms: Adaptation current time constant.

a::FT = 4nS: Subthreshold adaptation conductance.

b::FT = 80.5pA: Spike-triggered adaptation increment.

AP_membrane::FT = 10.0f0mV: After-potential membrane parameter .

BAP::FT = 1.0f0mV: Backpropagating action potential parameter.

up::IT = 1ms, τabs::IT = 2ms: Parameters related to spikes.

The types FT and IT represent Float32 and Int64 respectively.

SpikingNeuralNetworks.AdditiveNorm — Type

AdditiveNorm{FT = Float32} <: NormParam

This struct holds the parameters for additive normalization. It includes a timescale τ (default 0.0) and an operator (default addition).

SpikingNeuralNetworks.Dendrite — Type

Dendrite

A structure representing a dendritic compartment within a neuron model.

Fields

Er::FT = -70.6mV: Resting potential.

C::FT = 10pF: Membrane capacitance.

gax::FT = 10nS: Axial conductance.

gm::FT = 1nS: Membrane conductance.

l::FT = 150um: Length of the dendritic compartment.

d::FT = 4um: Diameter of the dendrite.

The type FT represents Float32.

SpikingNeuralNetworks.EmptyParam — Type

EmptyParam

A struct representing an empty parameter.

Fields

type::Symbol: The type of the parameter, default is :empty.

SpikingNeuralNetworks.GABAergic — Type

GABAergic struct represents a group of GABAergic receptors.

Fields

GABAa::T: GABAa receptor

GABAb::T: GABAb receptor

SpikingNeuralNetworks.Glutamatergic — Type

Glutamatergic struct represents a group of glutamatergic receptors.

Fields

AMPA::T: AMPA receptor

NMDA::T: NMDA receptor

SpikingNeuralNetworks.MultiplicativeNorm — Type

MultiplicativeNorm{FT = Int32} <: NormParam

This struct holds the parameters for multiplicative normalization. It includes a timescale τ (default 0.0) and an operator (default multiplication).

SpikingNeuralNetworks.NMDAVoltageDependency — Type

NMDAVoltageDependency struct represents the voltage dependence of NMDA receptors.

Fields

b::T: Voltage dependence factor (default: 3.36)

k::T: Voltage dependence factor (default: -0.077)

mg::T: Magnesium concentration (default: 1.0)

SpikingNeuralNetworks.NormParam — Type

Abstract type for normalization parameters.

SpikingNeuralNetworks.PlasticityVariables — Type

PlasticityVariables

An abstract type representing plasticity variables.

SpikingNeuralNetworks.PoissonStimulusFixed — Type

PoissonStimulusFixed

Poisson stimulus with fixed rate. The rate arrives to all the neuronstargeted
by the stimulus.

# Fields
- `rate::Vector{R}`: A vector containing the rate of the Poisson stimulus.
- `active::Vector{Bool}`: A vector of booleans indicating if the stimulus is active.

SpikingNeuralNetworks.PoissonStimulusInterval — Type

PoissonStimulusInterval

Poisson stimulus with rate defined for each cell in the layer. Each neuron of the 'N' Poisson population fires with 'rate' in the intervals defined by 'intervals'.

# Fields
- `rate::Vector{R}`: A vector containing the rate of the Poisson stimulus.
- `intervals::Vector{Vector{R}}`: A vector of vectors containing the intervals in which the Poisson stimulus is active.
- `active::Vector{Bool}`: A vector of booleans indicating if the stimulus is active.

SpikingNeuralNetworks.PoissonStimulusLayer — Type

PoissonStimulusLayer

Poisson stimulus with rate defined for each cell in the layer. Each neuron of the 'N' Poisson population fires with 'rate'.
The connectivity is defined by the parameter 'ϵ'. Thus, the number of presynaptic neuronsconnected to the postsynaptic neuronsis 'N*ϵ'. Each post-synaptic cell receives rate: 'rate * N * ϵ'.

# Fields
- `rate::Vector{R}`: A vector containing the rate of the Poisson stimulus.
- `N::Int32`: The number of neuronsin the layer.
- `ϵ::Float32`: The fraction of presynaptic neuronsconnected to the postsynaptic neurons.
- `active::Vector{Bool}`: A vector of booleans indicating if the stimulus is active.

SpikingNeuralNetworks.PoissonStimulusVariable — Type

PoissonStimulusVariable

Poisson stimulus with rate defined with a function.

# Fields
- `variables::Dict{Symbol,Any}`: A dictionary containing the variables for the function.
- `rate::Function`: A function defining the rate of the Poisson stimulus.
- `active::Vector{Bool}`: A vector of booleans indicating if the stimulus is active.

SpikingNeuralNetworks.PostSpike — Type

PostSpike

A structure defining the parameters of a post-synaptic spike event.

Fields

A::FT: Amplitude of the Post-Synaptic Potential (PSP).

τA::FT: Time constant of the PSP.

The type FT represents Float32.

SpikingNeuralNetworks.Receptor — Type

Receptor struct represents a synaptic receptor with parameters for reversal potential, rise time, decay time, and conductance.

Fields

E_rev::T: Reversal potential (default: 0.0)

τr::T: Rise time constant (default: -1.0)

τd::T: Decay time constant (default: -1.0)

g0::T: Maximum conductance (default: 0.0)

gsyn::T: Synaptic conductance (default: calculated based on g0, τr, and τd)

α::T: Alpha factor for the differential equation (default: calculated based on τr and τd)

τr⁻::T: Inverse of rise time constant (default: calculated based on τr)

τd⁻::T: Inverse of decay time constant (default: calculated based on τd)

nmda::T: NMDA factor (default: 0.0)

SpikingNeuralNetworks.Spiketimes — Type

Spiketimes

A type alias for a vector of vectors of Float32, representing spike times.

SpikingNeuralNetworks.Synapse — Type

Synapse struct represents a synaptic connection with different types of receptors.

Fields

AMPA::T: AMPA receptor

NMDA::T: NMDA receptor

GABAa::T: GABAa receptor

GABAb::T: GABAb receptor

SpikingNeuralNetworks.Synapse — Method

Construct a Synapse from Glutamatergic and GABAergic receptors.

Arguments

glu::Glutamatergic: Glutamatergic receptors

gaba::GABAergic: GABAergic receptors

Returns

Synapse: A Synapse object

SpikingNeuralNetworks.Time — Type

struct Time
Time

A mutable struct representing time. A mutable struct representing time.

Fields

t::Vector{Float32}: A vector containing the current time.

tt::Vector{Int}: A vector containing the current time step.

dt::Float32: The time step size.

Helper macros

SpikingNeuralNetworks.@load_units — Macro

@load_units
Load all the units defined in the module into the current scope.
This macro generates a block of expressions that assign the unit constants
    
The base units in the module are:
- cm : centimeters
- ms : milliseconds
- kHz : kilohertz
- mV : millivolts
- pA : picoamperes
- pF : picofarads
- nS : nanosiemens
- GΩ : gigaohms
- uM : micromolar

The derived units in the module are obtained as multiple or division of the base units. 

The standard integration time step is 0.125ms, which is used in the simulation.

SpikingNeuralNetworks.@snn_kw — Macro

A minimal implementation of Base.@kwdef with default type parameter support